Add socket activation to agent and server
This adds socket activation to spire.
This is useful because socket activation can be used
for implicit ordering between services as it makes
the socket available before the service starts.

+++ pkg/agent/api/endpoints_posix.go
@@ -11,6 +11,11 @@ import (
 )
 
 func (e *Endpoints) createListener() (net.Listener, error) {
+	// socket activation
+	if e.c.Listener != nil {
+		return e.listener.WrapListener(e.c.Listener)
+	}
+
 	// Remove uds if already exists
 	os.Remove(e.c.BindAddr.String())
